refactor(base): move internals to `internal` dir (#7770)
