+++ .github/workflows/test.yml
@@ -41,8 +41,13 @@ jobs:
           yarn add "@types/react@18" "@types/react-dom@18" --dev
           yarn add react@18 react-dom@18
 
+      - uses: actions/download-artifact@634f93cb2916e3fdff6788551b99b062d0335ce0 # v5.0.0
+        with:
+          name: build-${{ matrix.react }}
+          path: packages
+
       - name: Cypress run
-        uses: cypress-io/github-action@v6
+        uses: cypress-io/github-action@b8ba51a856ba5f4c15cf39007636d4ab04f23e3c # v6.10.2
         with:
           browser: chrome
           component: true
