+++ REUSE.toml
@@ -16,19 +16,19 @@ SPDX-FileCopyrightText = "2016-2025 Recharts Group"
 SPDX-License-Identifier = "MIT"
 
 [[annotations]]
-path = "packages/base/src/hooks/useIsomorphicLayoutEffect.ts"
+path = "packages/base/src/internal/hooks/useIsomorphicLayoutEffect.ts"
 precedence = "aggregate"
 SPDX-FileCopyrightText = "2019 Tanner Linsley"
 SPDX-License-Identifier = "MIT"
 
 [[annotations]]
-path = "packages/base/src/utils/debounce.ts"
+path = "packages/base/src/internal/utils/debounce.ts"
 precedence = "aggregate"
 SPDX-FileCopyrightText = "2009-2021 Jeremy Ashkenas, Julian Gonggrijp, and DocumentCloud and Investigative Reporters & Editors"
 SPDX-License-Identifier = "MIT"
 
 [[annotations]]
-path = "packages/base/src/utils/throttle.ts"
+path = "packages/base/src/internal/utils/throttle.ts"
 precedence = "aggregate"
 SPDX-FileCopyrightText = "2009-2021 Jeremy Ashkenas, Julian Gonggrijp, and DocumentCloud and Investigative Reporters & Editors"
 SPDX-License-Identifier = "MIT"
