+++ packages/base/src/hooks/index.ts
@@ -1,17 +1,4 @@
-import { useCurrentTheme } from './useCurrentTheme.js';
 import { useI18nBundle } from './useI18nBundle.js';
-import { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect.js';
-import { useIsRTL } from './useIsRTL.js';
-import { useStylesheet } from './useStylesheet.js';
-import { useSyncRef } from './useSyncRef.js';
 import { useViewportRange } from './useViewportRange.js';
 
-export {
-  useI18nBundle,
-  useIsomorphicLayoutEffect,
-  useIsRTL,
-  useSyncRef,
-  useViewportRange,
-  useStylesheet,
-  useCurrentTheme,
-};
+export { useI18nBundle, useViewportRange };
