+++ packages/base/src/internal/types/Ui5CustomEvent.ts
@@ -1,3 +1,6 @@
+/**
+ * ⚠️ __INTERNAL__ use only! This interface is not part of the public API.
+ */
 export interface Ui5CustomEvent<EventTarget = HTMLElement, Detail = never>
   extends Omit<CustomEvent<Detail>, 'target' | 'currentTarget'> {
   target: EventTarget;
