+++ packages/base/src/internal/utils/addCustomCSSWithScoping.ts
