+++ packages/compat/src/components/TableColumn/index.tsx
@@ -4,7 +4,7 @@ import '@ui5/webcomponents-compat/dist/TableColumn.js';
 import type TableColumnPopinDisplay from '@ui5/webcomponents-compat/dist/types/TableColumnPopinDisplay.js';
 import { getCompatCustomElementsScopingSuffix } from '@ui5/webcomponents-compat/dist/utils/CompatCustomElementsScope.js';
 import type { CommonProps, Ui5DomRef } from '@ui5/webcomponents-react-base';
-import { withWebComponent } from '@ui5/webcomponents-react-base/dist/wrapper/withWebComponent.js';
+import { withWebComponent } from '@ui5/webcomponents-react-base/withWebComponent';
 import type { ReactNode } from 'react';
 
 const compatScopingSuffix = getCompatCustomElementsScopingSuffix();
