+++ tsconfig.spec.json
@@ -3,7 +3,7 @@
     "target": "ES2021",
     "lib": ["es2022", "dom"],
     "types": ["cypress", "node", "@testing-library/cypress", "cypress-real-events", "chai"],
-    "moduleResolution": "Node",
+    "moduleResolution": "bundler",
     "jsx": "react-jsx",
     "composite": true
   },
